Split slow test files in half
